backend/requirements.txt
flask
flask-cors
supabase
google-generativeai
python-dotenv
